Small OS X changes: Change some ints to NSIntegers for x86_64. Update the nib. Throw errors when building clip_macosx.cpp (32-bit Carbon) on 64-bit machines. Update the PrefsEditor.

+++ SheepShaver/src/MacOSX/Launcher/VMSettingsController.mm
@@ -60,12 +60,12 @@ void prefs_exit()
   return self;
 }
 
-- (int) numberOfRowsInTableView: (NSTableView *) table
+- (NSInteger) numberOfRowsInTableView: (NSTableView *) table
 {
   return [diskArray count];
 }
 
-- (id) tableView: (NSTableView *) table objectValueForTableColumn: (NSTableColumn *) col row: (int) row
+- (id) tableView: (NSTableView *) table objectValueForTableColumn: (NSTableColumn *) col row: (NSInteger) row
 {
   return [diskArray objectAtIndex: row];
 }
